Cleaned up handlers and cloning logging
