+++ internal/api/handlers/auth_handler.go
@@ -6,14 +6,14 @@ import (
 	"net/http"
 
 	"github.com/cpp-cyber/proclone/internal/auth"
+	"github.com/cpp-cyber/proclone/internal/proxmox"
 	"github.com/gin-contrib/sessions"
 	"github.com/gin-gonic/gin"
 )
 
-// AuthHandler handles HTTP authentication requests
-type AuthHandler struct {
-	authService auth.Service
-}
+// =================================================
+// Login / Logout / Session Handlers
+// =================================================
 
 // NewAuthHandler creates a new authentication handler
 func NewAuthHandler() (*AuthHandler, error) {
@@ -22,29 +22,30 @@ func NewAuthHandler() (*AuthHandler, error) {
 		return nil, fmt.Errorf("failed to create auth service: %w", err)
 	}
 
+	proxmoxService, err := proxmox.NewService()
+	if err != nil {
+		return nil, fmt.Errorf("failed to create proxmox service: %w", err)
+	}
+
 	log.Println("Auth handler initialized")
 
 	return &AuthHandler{
-		authService: authService,
+		authService:    authService,
+		proxmoxService: proxmoxService,
 	}, nil
 }
 
 // LoginHandler handles the login POST request
 func (h *AuthHandler) LoginHandler(c *gin.Context) {
-	var loginReq struct {
-		Username string `json:"username" binding:"required"`
-		Password string `json:"password" binding:"required"`
-	}
-
-	if err := c.ShouldBindJSON(&loginReq); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request format"})
+	var req UserRequest
+	if !ValidateAndBind(c, &req) {
 		return
 	}
 
 	// Authenticate user
-	valid, err := h.authService.Authenticate(loginReq.Username, loginReq.Password)
+	valid, err := h.authService.Authenticate(req.Username, req.Password)
 	if err != nil {
-		log.Printf("Authentication failed for user %s: %v", loginReq.Username, err)
+		log.Printf("Authentication failed for user %s: %v", req.Username, err)
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "Authentication failed"})
 		return
 	}
@@ -56,18 +57,18 @@ func (h *AuthHandler) LoginHandler(c *gin.Context) {
 
 	// Create session
 	session := sessions.Default(c)
-	session.Set("id", loginReq.Username)
+	session.Set("id", req.Username)
 
 	// Check if user is admin
-	isAdmin, err := h.authService.IsAdmin(loginReq.Username)
+	isAdmin, err := h.authService.IsAdmin(req.Username)
 	if err != nil {
-		log.Printf("Error checking admin status for user %s: %v", loginReq.Username, err)
+		log.Printf("Error checking admin status for user %s: %v", req.Username, err)
 		isAdmin = false
 	}
 	session.Set("isAdmin", isAdmin)
 
 	if err := session.Save(); err != nil {
-		log.Printf("Failed to save session for user %s: %v", loginReq.Username, err)
+		log.Printf("Failed to save session for user %s: %v", req.Username, err)
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to save session"})
 		return
 	}
@@ -114,9 +115,8 @@ func (h *AuthHandler) SessionHandler(c *gin.Context) {
 }
 
 func (h *AuthHandler) RegisterHandler(c *gin.Context) {
-	var req CreateUserRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request format"})
+	var req UserRequest
+	if !ValidateAndBind(c, &req) {
 		return
 	}
 
@@ -139,4 +139,274 @@ func (h *AuthHandler) RegisterHandler(c *gin.Context) {
 	}
 
 	c.JSON(http.StatusOK, gin.H{"message": "User registered successfully"})
+}
+
+// =================================================
+// User Handlers
+// =================================================
+
+// ADMIN: GetUsersHandler returns a list of all users
+func (h *AuthHandler) GetUsersHandler(c *gin.Context) {
+	users, err := h.authService.GetUsers()
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to retrieve users"})
+		return
+	}
+
+	var adminCount = 0
+	var disabledCount = 0
+	for _, user := range users {
+		if user.IsAdmin {
+			adminCount++
+		}
+		if !user.Enabled {
+			disabledCount++
+		}
+	}
+
+	c.JSON(http.StatusOK, gin.H{
+		"users":          users,
+		"count":          len(users),
+		"disabled_count": disabledCount,
+		"admin_count":    adminCount,
+	})
+}
+
+// ADMIN: CreateUsersHandler creates new user(s)
+func (h *AuthHandler) CreateUsersHandler(c *gin.Context) {
+	var req AdminCreateUserRequest
+	if !ValidateAndBind(c, &req) {
+		return
+	}
+
+	var errors []error
+
+	// Create users in AD
+	for _, user := range req.Users {
+		if err := h.authService.CreateAndRegisterUser(auth.UserRegistrationInfo(user)); err != nil {
+			errors = append(errors, fmt.Errorf("failed to create user %s: %v", user.Username, err))
+		}
+	}
+
+	if len(errors) > 0 {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create users", "details": errors})
+		return
+	}
+
+	// Sync users to Proxmox
+	if err := h.proxmoxService.SyncUsers(); err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to sync users with Proxmox", "details": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusCreated, gin.H{"message": "Users created successfully"})
+}
+
+// ADMIN: DeleteUsersHandler deletes existing user(s)
+func (h *AuthHandler) DeleteUsersHandler(c *gin.Context) {
+	var req UsersRequest
+	if !ValidateAndBind(c, &req) {
+		return
+	}
+
+	var errors []error
+
+	// Delete users in AD
+	for _, username := range req.Usernames {
+		if err := h.authService.DeleteUser(username); err != nil {
+			errors = append(errors, fmt.Errorf("failed to delete user %s: %v", username, err))
+		}
+	}
+
+	if len(errors) > 0 {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete users", "details": errors})
+		return
+	}
+
+	// Sync users to Proxmox
+	if err := h.proxmoxService.SyncUsers(); err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to sync users with Proxmox", "details": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Users deleted successfully"})
+}
+
+// ADMIN: EnableUsersHandler enables existing user(s)
+func (h *AuthHandler) EnableUsersHandler(c *gin.Context) {
+	var req UsersRequest
+	if !ValidateAndBind(c, &req) {
+		return
+	}
+
+	var errors []error
+
+	for _, username := range req.Usernames {
+		if err := h.authService.EnableUserAccount(username); err != nil {
+			errors = append(errors, fmt.Errorf("failed to enable user %s: %v", username, err))
+		}
+	}
+
+	if len(errors) > 0 {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to enable users", "details": errors})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Users enabled successfully"})
+}
+
+// ADMIN: DisableUsersHandler disables existing user(s)
+func (h *AuthHandler) DisableUsersHandler(c *gin.Context) {
+	var req UsersRequest
+	if !ValidateAndBind(c, &req) {
+		return
+	}
+
+	var errors []error
+
+	for _, username := range req.Usernames {
+		if err := h.authService.DisableUserAccount(username); err != nil {
+			errors = append(errors, fmt.Errorf("failed to disable user %s: %v", username, err))
+		}
+	}
+
+	if len(errors) > 0 {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to disable users", "details": errors})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Users disabled successfully"})
+}
+
+// =================================================
+// Group Functions
+// =================================================
+
+// ADMIN: SetUserGroupsHandler sets the groups for an existing user
+func (h *AuthHandler) SetUserGroupsHandler(c *gin.Context) {
+	var req SetUserGroupsRequest
+	if !ValidateAndBind(c, &req) {
+		return
+	}
+
+	if err := h.authService.SetUserGroups(req.Username, req.Groups); err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to set user groups", "details": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "User groups updated successfully"})
+}
+
+func (h *AuthHandler) GetGroupsHandler(c *gin.Context) {
+	groups, err := h.authService.GetGroups()
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to retrieve groups"})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{
+		"groups": groups,
+		"count":  len(groups),
+	})
+}
+
+// ADMIN: CreateGroupsHandler creates new group(s)
+func (h *AuthHandler) CreateGroupsHandler(c *gin.Context) {
+	var req GroupsRequest
+	if !ValidateAndBind(c, &req) {
+		return
+	}
+
+	var errors []error
+
+	// Create groups in AD
+	for _, group := range req.Groups {
+		if err := h.authService.CreateGroup(group); err != nil {
+			errors = append(errors, fmt.Errorf("failed to create group %s: %v", group, err))
+		}
+	}
+
+	if len(errors) > 0 {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create groups", "details": errors})
+		return
+	}
+
+	// Sync groups to Proxmox
+	if err := h.proxmoxService.SyncGroups(); err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to sync groups with Proxmox", "details": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusCreated, gin.H{"message": "Groups created successfully"})
+}
+
+func (h *AuthHandler) RenameGroupHandler(c *gin.Context) {
+	var req RenameGroupRequest
+	if !ValidateAndBind(c, &req) {
+		return
+	}
+
+	if err := h.authService.RenameGroup(req.OldName, req.NewName); err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to rename group"})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Group renamed successfully"})
+}
+
+func (h *AuthHandler) DeleteGroupsHandler(c *gin.Context) {
+	var req GroupsRequest
+	if !ValidateAndBind(c, &req) {
+		return
+	}
+
+	var errors []error
+
+	// Delete groups in AD
+	for _, group := range req.Groups {
+		if err := h.authService.DeleteGroup(group); err != nil {
+			errors = append(errors, fmt.Errorf("failed to delete group %s: %v", group, err))
+		}
+	}
+
+	if len(errors) > 0 {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete groups", "details": errors})
+		return
+	}
+
+	// Sync groups to Proxmox
+	if err := h.proxmoxService.SyncGroups(); err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to sync groups with Proxmox", "details": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Groups deleted successfully"})
+}
+
+func (h *AuthHandler) AddUsersHandler(c *gin.Context) {
+	var req ModifyGroupMembersRequest
+	if !ValidateAndBind(c, &req) {
+		return
+	}
+
+	if err := h.authService.AddUsersToGroup(req.Group, req.Usernames); err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to add users to group"})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Users added to group successfully"})
+}
+
+func (h *AuthHandler) RemoveUsersHandler(c *gin.Context) {
+	var req ModifyGroupMembersRequest
+	if !ValidateAndBind(c, &req) {
+		return
+	}
+
+	if err := h.authService.RemoveUsersFromGroup(req.Group, req.Usernames); err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to remove users from group"})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Users removed from group successfully"})
 }
